wip, restructure func names and push apigateway endpoint to ssm
