+++ serverless/functions/package.json
@@ -1,7 +1,10 @@
 {
   "devDependencies": {
+    "serverless-domain-manager": "^8.0.0",
+    "serverless-export-env": "^2.2.0",
     "serverless-python-requirements": "^6.1.2",
     "serverless-ssm-fetch": "^2.0.0",
-    "serverless-stack-output": "^0.2.3"
+    "serverless-stack-output": "^0.2.3",
+    "true": "^0.0.4"
   }
 }
